@@ -13,9 +13,9 @@ var paths = {
     js: webroot + "js/**/*.js",
     minJs: webroot + "js/**/*.min.js",
     css: webroot + "css/**/*.css",
-    minCss: webroot + "css/**/*.min.css",
-    concatJsDest: webroot + "js/site.min.js",
-    concatCssDest: webroot + "css/site.min.css"
+    minCss: webroot + "dist/css/**/*.min.css",
+    concatJsDest: webroot + "dist/js/site.min.js",
+    concatCssDest: webroot + "dist/css/site.min.css"
 };
 
 gulp.task("clean:js", function (cb) {
@@ -26,7 +26,7 @@ gulp.task("clean:css", function (cb) {
     rimraf(paths.concatCssDest, cb);
 });
 
-gulp.task("clean", ["clean:js", "clean:css"]);
+gulp.task("clean", gulp.series("clean:js", "clean:css"));
 
 gulp.task("min:js", function () {
     return gulp.src([paths.js, "!" + paths.minJs], { base: "." })
@@ -42,4 +42,4 @@ gulp.task("min:css", function () {
         .pipe(gulp.dest("."));
 });
 
-gulp.task("min", ["min:js", "min:css"]);
+gulp.task("min", gulp.series("min:js", "min:css"));
